refactor(theme): remove all themes
