+++ packages/yoga/src/Theme/index.native.js
@@ -1,5 +1,5 @@
 import ThemeProvider from './Provider';
-import * as themes from './themes';
+import yogaTheme from './theme';
 import theme from './helpers/themeReader';
 
-export { ThemeProvider as default, themes, theme };
+export { ThemeProvider as default, yogaTheme, theme };
