+++ packages/yoga/src/Theme/theme/index.js
@@ -1,0 +1,3 @@
+import theme from './theme';
+
+export default theme;
