--- packages/yoga/src/Theme/themes/EndUser.js
@@ -1,16 +1,0 @@
-import BaseTheme from './BaseTheme';
-
-const EndUser = tokens => {
-  const colors = {
-    primary: tokens.colors.madrid,
-    secondary: tokens.colors.paris,
-    tertiary: tokens.colors.saoPaulo,
-  };
-
-  return {
-    colors,
-    ...BaseTheme(colors),
-  };
-};
-
-export default EndUser;
