--- packages/yoga/src/Theme/themes/index.js
@@ -1,7 +1,0 @@
-import Corporate from './Corporate';
-import EndUser from './EndUser';
-import Gyms from './Gyms';
-import Wellness from './Wellness';
-import Kids from './Kids';
-
-export { EndUser as default, EndUser, Corporate, Gyms, Wellness, Kids };
